+++ pkg/metrics/instance/instance.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
+	"github.com/grafana/agent/internal/agentseed"
 	"github.com/grafana/agent/internal/useragent"
 	"github.com/grafana/agent/pkg/metrics/wal"
 	"github.com/grafana/agent/pkg/util"
@@ -157,7 +158,6 @@ func (c *Config) ApplyDefaults(global GlobalConfig) error {
 	}
 
 	rwNames := map[string]struct{}{}
-
 	// If the instance remote write is not filled in, then apply the prometheus write config
 	if len(c.RemoteWrite) == 0 {
 		c.RemoteWrite = c.global.RemoteWrite
@@ -166,7 +166,6 @@ func (c *Config) ApplyDefaults(global GlobalConfig) error {
 		if cfg == nil {
 			return fmt.Errorf("empty or null remote write config section")
 		}
-
 		// Typically Prometheus ignores empty names here, but we need to assign a
 		// unique name to the config so we can pull metrics from it when running
 		// an instance.
@@ -183,7 +182,6 @@ func (c *Config) ApplyDefaults(global GlobalConfig) error {
 			cfg.Name = c.Name + "-" + hash[:6]
 			generatedName = true
 		}
-
 		if _, exists := rwNames[cfg.Name]; exists {
 			if generatedName {
 				return fmt.Errorf("found two identical remote_write configs")
@@ -419,6 +417,13 @@ func (i *Instance) initialize(ctx context.Context, reg prometheus.Registerer, cf
 	// Set up the remote storage
 	remoteLogger := log.With(i.logger, "component", "remote")
 	i.remoteStore = remote.NewStorage(remoteLogger, reg, i.wal.StartTime, i.wal.Directory(), cfg.RemoteFlushDeadline, i.readyScrapeManager)
+	uid := agentseed.Get().UID
+	for _, rw := range cfg.RemoteWrite {
+		if rw.Headers == nil {
+			rw.Headers = map[string]string{}
+		}
+		rw.Headers[agentseed.HeaderName] = uid
+	}
 	err = i.remoteStore.ApplyConfig(&config.Config{
 		GlobalConfig:       cfg.global.Prometheus,
 		RemoteWriteConfigs: cfg.RemoteWrite,